refactor: ライブラリモジュールをリネーム
