+++ test/test_library_manager.py
@@ -11,7 +11,7 @@ from zipfile import ZipFile
 
 from fastapi import HTTPException
 
-from voicevox_engine.library_manager import LibraryManager
+from voicevox_engine.library import LibraryManager
 
 vvlib_manifest_name = "vvlib_manifest.json"
