+++ voicevox_engine/app/application.py
@@ -18,7 +18,7 @@ from voicevox_engine.app.routers.user_dict import generate_user_dict_router
 from voicevox_engine.cancellable_engine import CancellableEngine
 from voicevox_engine.core.core_adapter import CoreAdapter
 from voicevox_engine.engine_manifest.EngineManifestLoader import load_manifest
-from voicevox_engine.library_manager import LibraryManager
+from voicevox_engine.library import LibraryManager
 from voicevox_engine.metas.MetasStore import MetasStore
 from voicevox_engine.preset.PresetManager import PresetManager
 from voicevox_engine.setting.Setting import CorsPolicyMode
